fix: :bug: missing Branch and Service foreign keys have been added!

+++ backend/src/main/java/com/enterprisemanager/backend/domain/entities/ServiceBranch.java
@@ -24,5 +24,5 @@ public class ServiceBranch {
     private Branch branches;
 
     @Column(columnDefinition = "DECIMAL(16,2)",nullable = false)
-    private Double serviceValue;
+    private Float serviceValue;
 }
